+++ playwright/fixtures.ts
@@ -9,6 +9,8 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
     // Authenticate once per worker with a worker-scoped fixture.
     workerStorageState: [
         async ({ browser }, use) => {
+            console.log('Authenticating...');
+
             // Use parallelIndex as a unique identifier for each worker.
             const id = test.info().parallelIndex;
             const fileName = path.resolve('playwright', '.auth', `${id}.json`);
@@ -20,7 +22,6 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
             }
 
             // Important: make sure we authenticate in a clean environment by unsetting storage state.
-            // @ts-ignore It accepts undefined but types are wrong.
             const page = await browser.newPage({
                 baseURL: 'http://localhost:4173',
             });
@@ -44,6 +45,17 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
                 .getByTestId('password-repeat-field')
                 .fill(account.password);
             await page.getByTestId('join-button').click();
+
+            const url = await page.evaluate(() => document.location.href);
+
+            console.log(
+                'Filled in ' +
+                    account.username +
+                    ' and ' +
+                    account.password +
+                    ' and arrived at ' +
+                    url,
+            );
 
             // Wait for the final redirect URL to ensure that the cookies are actually set.
             await page.waitForURL('/profile');